requirements.txt
Flask==2.0.2
Werkzeug==2.0.0
ipaddress==1.0.23
MarkupSafe==2.1.1
python-dotenv==0.19.2
gunicorn==20.1.0